Move tests out of AbstractDatabaseAdapterTest
