+++ versioned/persist/tests/src/main/java/org/projectnessie/versioned/persist/tests/AbstractReferences.java
@@ -1,0 +1,247 @@
+/*
+ * Copyright (C) 2020 Dremio
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.projectnessie.versioned.persist.tests;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+import com.google.protobuf.ByteString;
+import java.util.Arrays;
+import java.util.Optional;
+import java.util.stream.Stream;
+import org.junit.jupiter.api.Test;
+import org.projectnessie.versioned.BranchName;
+import org.projectnessie.versioned.GetNamedRefsParams;
+import org.projectnessie.versioned.Hash;
+import org.projectnessie.versioned.Key;
+import org.projectnessie.versioned.NamedRef;
+import org.projectnessie.versioned.ReferenceAlreadyExistsException;
+import org.projectnessie.versioned.ReferenceConflictException;
+import org.projectnessie.versioned.ReferenceInfo;
+import org.projectnessie.versioned.ReferenceNotFoundException;
+import org.projectnessie.versioned.TagName;
+import org.projectnessie.versioned.persist.adapter.ContentId;
+import org.projectnessie.versioned.persist.adapter.DatabaseAdapter;
+import org.projectnessie.versioned.persist.adapter.ImmutableCommitParams;
+import org.projectnessie.versioned.persist.adapter.KeyWithBytes;
+
+/** Verifies handling of repo-description in the database-adapters. */
+public abstract class AbstractReferences {
+
+  private final DatabaseAdapter databaseAdapter;
+
+  protected AbstractReferences(DatabaseAdapter databaseAdapter) {
+    this.databaseAdapter = databaseAdapter;
+  }
+
+  @Test
+  void createBranch() throws Exception {
+    BranchName create = BranchName.of("createBranch");
+    createNamedRef(create, TagName.of(create.getName()));
+  }
+
+  @Test
+  void createTag() throws Exception {
+    TagName create = TagName.of("createTag");
+    createNamedRef(create, BranchName.of(create.getName()));
+  }
+
+  private void createNamedRef(NamedRef create, NamedRef opposite) throws Exception {
+    BranchName branch = BranchName.of("main");
+
+    try (Stream<ReferenceInfo<ByteString>> refs =
+        databaseAdapter.namedRefs(GetNamedRefsParams.DEFAULT)) {
+      assertThat(refs.map(ReferenceInfo::getNamedRef)).containsExactlyInAnyOrder(branch);
+    }
+
+    Hash mainHash = databaseAdapter.hashOnReference(branch, Optional.empty());
+
+    assertThatThrownBy(() -> databaseAdapter.hashOnReference(create, Optional.empty()))
+        .isInstanceOf(ReferenceNotFoundException.class);
+
+    Hash createHash =
+        databaseAdapter.create(create, databaseAdapter.hashOnReference(branch, Optional.empty()));
+    assertThat(createHash).isEqualTo(mainHash);
+
+    try (Stream<ReferenceInfo<ByteString>> refs =
+        databaseAdapter.namedRefs(GetNamedRefsParams.DEFAULT)) {
+      assertThat(refs.map(ReferenceInfo::getNamedRef)).containsExactlyInAnyOrder(branch, create);
+    }
+
+    assertThatThrownBy(
+            () ->
+                databaseAdapter.create(
+                    create, databaseAdapter.hashOnReference(branch, Optional.empty())))
+        .isInstanceOf(ReferenceAlreadyExistsException.class);
+
+    assertThat(databaseAdapter.hashOnReference(create, Optional.empty())).isEqualTo(createHash);
+    assertThatThrownBy(() -> databaseAdapter.hashOnReference(opposite, Optional.empty()))
+        .isInstanceOf(ReferenceNotFoundException.class);
+
+    assertThatThrownBy(
+            () ->
+                databaseAdapter.create(
+                    BranchName.of(create.getName()),
+                    databaseAdapter.hashOnReference(branch, Optional.empty())))
+        .isInstanceOf(ReferenceAlreadyExistsException.class);
+
+    assertThatThrownBy(
+            () -> databaseAdapter.delete(create, Optional.of(Hash.of("dead00004242fee18eef"))))
+        .isInstanceOf(ReferenceConflictException.class);
+
+    assertThatThrownBy(() -> databaseAdapter.delete(opposite, Optional.of(createHash)))
+        .isInstanceOf(ReferenceNotFoundException.class);
+
+    databaseAdapter.delete(create, Optional.of(createHash));
+
+    assertThatThrownBy(() -> databaseAdapter.hashOnReference(create, Optional.empty()))
+        .isInstanceOf(ReferenceNotFoundException.class);
+
+    try (Stream<ReferenceInfo<ByteString>> refs =
+        databaseAdapter.namedRefs(GetNamedRefsParams.DEFAULT)) {
+      assertThat(refs.map(ReferenceInfo::getNamedRef)).containsExactlyInAnyOrder(branch);
+    }
+  }
+
+  @Test
+  void verifyNotFoundAndConflictExceptionsForUnreachableCommit() throws Exception {
+    BranchName main = BranchName.of("main");
+    BranchName unreachable = BranchName.of("unreachable");
+    BranchName helper = BranchName.of("helper");
+
+    databaseAdapter.create(unreachable, databaseAdapter.hashOnReference(main, Optional.empty()));
+    Hash helperHead =
+        databaseAdapter.create(helper, databaseAdapter.hashOnReference(main, Optional.empty()));
+
+    Hash unreachableHead =
+        databaseAdapter.commit(
+            ImmutableCommitParams.builder()
+                .toBranch(unreachable)
+                .commitMetaSerialized(ByteString.copyFromUtf8("commit meta"))
+                .addPuts(
+                    KeyWithBytes.of(
+                        Key.of("foo"),
+                        ContentId.of("contentId"),
+                        (byte) 0,
+                        ByteString.copyFromUtf8("hello")))
+                .build());
+
+    assertAll(
+        () ->
+            assertThatThrownBy(
+                    () -> databaseAdapter.hashOnReference(main, Optional.of(unreachableHead)))
+                .isInstanceOf(ReferenceNotFoundException.class)
+                .hasMessage(
+                    String.format(
+                        "Could not find commit '%s' in reference '%s'.",
+                        unreachableHead.asString(), main.getName())),
+        () ->
+            assertThatThrownBy(
+                    () ->
+                        databaseAdapter.commit(
+                            ImmutableCommitParams.builder()
+                                .toBranch(helper)
+                                .expectedHead(Optional.of(unreachableHead))
+                                .commitMetaSerialized(ByteString.copyFromUtf8("commit meta"))
+                                .addPuts(
+                                    KeyWithBytes.of(
+                                        Key.of("bar"),
+                                        ContentId.of("contentId-no-no"),
+                                        (byte) 0,
+                                        ByteString.copyFromUtf8("hello")))
+                                .build()))
+                .isInstanceOf(ReferenceNotFoundException.class)
+                .hasMessage(
+                    String.format(
+                        "Could not find commit '%s' in reference '%s'.",
+                        unreachableHead.asString(), helper.getName())),
+        () ->
+            assertThatThrownBy(
+                    () ->
+                        databaseAdapter.assign(
+                            helper,
+                            Optional.of(unreachableHead),
+                            databaseAdapter.hashOnReference(main, Optional.empty())))
+                .isInstanceOf(ReferenceConflictException.class)
+                .hasMessage(
+                    String.format(
+                        "Named-reference '%s' is not at expected hash '%s', but at '%s'.",
+                        helper.getName(), unreachableHead.asString(), helperHead.asString())));
+  }
+
+  @Test
+  void assign() throws Exception {
+    BranchName main = BranchName.of("main");
+    TagName tag = TagName.of("tag");
+    TagName branch = TagName.of("branch");
+
+    databaseAdapter.create(branch, databaseAdapter.hashOnReference(main, Optional.empty()));
+    databaseAdapter.create(tag, databaseAdapter.hashOnReference(main, Optional.empty()));
+
+    Hash beginning = databaseAdapter.hashOnReference(main, Optional.empty());
+
+    Hash[] commits = new Hash[3];
+    for (int i = 0; i < commits.length; i++) {
+      commits[i] =
+          databaseAdapter.commit(
+              ImmutableCommitParams.builder()
+                  .toBranch(main)
+                  .commitMetaSerialized(ByteString.copyFromUtf8("commit meta " + i))
+                  .addPuts(
+                      KeyWithBytes.of(
+                          Key.of("bar", Integer.toString(i)),
+                          ContentId.of("contentId-" + i),
+                          (byte) 0,
+                          ByteString.copyFromUtf8("hello " + i)))
+                  .build());
+    }
+
+    Hash expect = beginning;
+    for (Hash commit : commits) {
+      assertThat(
+              Arrays.asList(
+                  databaseAdapter.hashOnReference(branch, Optional.empty()),
+                  databaseAdapter.hashOnReference(tag, Optional.empty())))
+          .containsExactly(expect, expect);
+
+      databaseAdapter.assign(tag, Optional.of(expect), commit);
+
+      databaseAdapter.assign(branch, Optional.of(expect), commit);
+
+      expect = commit;
+    }
+
+    assertThat(
+            Arrays.asList(
+                databaseAdapter.hashOnReference(branch, Optional.empty()),
+                databaseAdapter.hashOnReference(tag, Optional.empty())))
+        .containsExactly(commits[commits.length - 1], commits[commits.length - 1]);
+  }
+
+  @Test
+  void recreateDefaultBranch() throws Exception {
+    BranchName main = BranchName.of("main");
+    Hash mainHead = databaseAdapter.hashOnReference(main, Optional.empty());
+    databaseAdapter.delete(main, Optional.of(mainHead));
+
+    assertThatThrownBy(() -> databaseAdapter.hashOnReference(main, Optional.empty()))
+        .isInstanceOf(ReferenceNotFoundException.class);
+
+    databaseAdapter.create(main, null);
+    databaseAdapter.hashOnReference(main, Optional.empty());
+  }
+}
